Fixed buffer overflow in printint (#299)
There is a buffer overflow in printint function which can be triggered if numbers with more than 16 digits are written to the terminal. Numbers 18446744073709551615 and -9223372036854775808 both have 20 digits for instance. To this extent, I increased the buffer size to 20 characters which can fit base 10 of all 64 bit numbers (negative or positive).